requirements.txt
# BioOracle Dependencies

# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.0

# Anthropic Claude API
anthropic==0.28.0

# HTTP and Async
aiohttp==3.9.5
httpx==0.27.0
tenacity==8.3.0
backoff==2.2.1

# Data Processing
pandas==2.2.2
numpy==1.26.4
lxml==5.2.2
xmltodict==0.13.0
openpyxl==3.1.4

# Storage
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.5
alembic==1.13.1

# Visualization
plotly==5.22.0
dash==2.17.1
dash-bootstrap-components==1.6.0
kaleido==0.2.1

# NLP / ML (for MeSH term processing)
scikit-learn==1.5.0
faiss-cpu==1.8.0

# Embeddings for RAG
sentence-transformers==3.0.1

# Email
secure-smtplib==0.1.1

# Task Queue
celery==5.4.0
flower==2.0.1

# Security and Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Export
reportlab==4.2.2
fpdf2==2.7.9
jinja2==3.1.4

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

# Logging
loguru==0.7.2

# Dev Tools
black==24.4.2
isort==5.13.2
flake8==7.0.0
